requirements.txt
camelot-py==0.11.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
et-xmlfile==1.1.0
ghostscript==0.7
numpy==1.25.0
opencv-python==4.8.0.74
openpyxl==3.1.2
pandas==2.0.3
pdfminer.six==20221105
pdftopng==0.2.3
pycparser==2.21
pypdf==3.12.0
pypdf2==2.12.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabulate==0.9.0
typing-extensions==4.7.1
tzdata==2023.3
